requirements.txt
requests
Pillow
pywal
colorthief
colorz
fast_colorthief
haishoku